requirements.txt
Flask==2.3.3
Flask-JWT-Extended==4.5.3
psycopg2==2.9.11
python-dotenv==1.0
Flask-CORS==4.0.0
bcrypt==4.0.1
openai==1.30.1
openpyxl==3.1.2
waitress==2.1.2